requirements.txt
# AutoAssignment 项目依赖
# 安装命令: pip install -r requirements.txt

# Word 文档处理
python-docx>=0.8.11

# HTTP 请求库
requests>=2.25.0
